ai-research-assistant/requirements.txt
langchain
openai
langchain-community
langchain-openrouter
langchain-openai
python-dotenv
streamlit
pydantic<2
duckduckgo-search
